refactor(renderer): Extract MaterialReport and MTKViewDelegate extensions
- Extract MaterialReport struct and generateMaterialReport() to VRMMaterialReport.swift
- Extract MTKViewDelegate conformance to VRMRenderer+MTKViewDelegate.swift
- Extract SpringBone debug controls to VRMRenderer+MTKViewDelegate.swift
- Reduces VRMRenderer.swift from 3362 to 3148 lines (-214 lines)
- Total reduction: 3504 -> 3148 lines (-356 lines, -10.2%)

Continues work on #4
