+++ Sources/VRMMetalKit/Renderer/VRMRenderer.swift
@@ -331,9 +331,9 @@ public final class VRMRenderer: NSObject, @unchecked Sendable {
     private var springBoneComputeSystem: SpringBoneComputeSystem?
     public var enableSpringBone: Bool = false
     private var lastUpdateTime: CFTimeInterval = 0
-    private var temporaryGravity: SIMD3<Float>?
-    private var temporaryWind: SIMD3<Float>?
-    private var forceTimer: Float = 0
+    var temporaryGravity: SIMD3<Float>?
+    var temporaryWind: SIMD3<Float>?
+    var forceTimer: Float = 0
 
     // OPTIMIZATION: Static zero weights array (avoids allocation per primitive)
     private static let zeroMorphWeights = [Float](repeating: 0, count: 8)
@@ -364,7 +364,7 @@ public final class VRMRenderer: NSObject, @unchecked Sendable {
     private var debugRenderer: VRMDebugRenderer?
 
     // Frame counter for debug logging
-    private var frameCounter = 0
+    var frameCounter = 0
 
     // PERFORMANCE OPTIMIZATION: Cached render items to avoid rebuilding every frame
     private var cachedRenderItems: [RenderItem]?
@@ -3144,219 +3144,5 @@ public final class VRMRenderer: NSObject, @unchecked Sendable {
         }
     }
 
-    // MARK: - Material Report Generation
-
-    public struct MaterialReport: Codable {
-        public let modelName: String
-        public let materials: [MaterialInfo]
-        public let summary: Summary
-
-        public struct MaterialInfo: Codable {
-            public let index: Int
-            public let name: String
-            public let alphaMode: String
-            public let alphaCutoff: Float
-            public let baseColorFactor: [Float]
-            public let hasBaseTexture: Bool
-            public let textureSize: [Int]?
-            public let doubleSided: Bool
-            public let mtoonShadeColor: [Float]?
-            public let hasAlphaIssue: Bool
-        }
-
-        public struct Summary: Codable {
-            public let totalMaterials: Int
-            public let opaqueCount: Int
-            public let maskCount: Int
-            public let blendCount: Int
-            public let suspiciousAlphaCount: Int
-        }
-    }
-
-    public func generateMaterialReport() -> MaterialReport? {
-        guard let model = model else {
-            vrmLog("[VRMRenderer] No model loaded for material report")
-            return nil
-        }
-
-        var materialInfos: [MaterialReport.MaterialInfo] = []
-        var opaqueCount = 0
-        var maskCount = 0
-        var blendCount = 0
-        var suspiciousAlphaCount = 0
-
-        for (index, material) in model.materials.enumerated() {
-            // Count alpha modes
-            switch material.alphaMode.lowercased() {
-            case "opaque":
-                opaqueCount += 1
-            case "mask":
-                maskCount += 1
-            case "blend":
-                blendCount += 1
-            default:
-                opaqueCount += 1
-            }
-
-            // Check for suspicious alpha values
-            let hasAlphaIssue = material.baseColorFactor.w < 0.01 ||
-                               (material.alphaMode.lowercased() == "opaque" && material.baseColorFactor.w < 1.0)
-            if hasAlphaIssue {
-                suspiciousAlphaCount += 1
-            }
-
-            // Get texture size if available
-            var textureSize: [Int]? = nil
-            if let baseTexture = material.baseColorTexture,
-               let mtlTexture = baseTexture.mtlTexture {
-                textureSize = [mtlTexture.width, mtlTexture.height]
-            }
-
-            // Get MToon shade color if available
-            var mtoonShadeColor: [Float]? = nil
-            if let mtoon = material.mtoon {
-                mtoonShadeColor = [
-                    mtoon.shadeColorFactor.x,
-                    mtoon.shadeColorFactor.y,
-                    mtoon.shadeColorFactor.z
-                ]
-            }
-
-            let info = MaterialReport.MaterialInfo(
-                index: index,
-                name: material.name ?? "Material_\(index)",
-                alphaMode: material.alphaMode,
-                alphaCutoff: material.alphaCutoff,
-                baseColorFactor: [
-                    material.baseColorFactor.x,
-                    material.baseColorFactor.y,
-                    material.baseColorFactor.z,
-                    material.baseColorFactor.w
-                ],
-                hasBaseTexture: material.baseColorTexture != nil,
-                textureSize: textureSize,
-                doubleSided: material.doubleSided,
-                mtoonShadeColor: mtoonShadeColor,
-                hasAlphaIssue: hasAlphaIssue
-            )
-
-            materialInfos.append(info)
-        }
-
-        let summary = MaterialReport.Summary(
-            totalMaterials: model.materials.count,
-            opaqueCount: opaqueCount,
-            maskCount: maskCount,
-            blendCount: blendCount,
-            suspiciousAlphaCount: suspiciousAlphaCount
-        )
-
-        return MaterialReport(
-            modelName: "VRM Model",  // VRMModel doesn't have a name property
-            materials: materialInfos,
-            summary: summary
-        )
-    }
-}
-
-// MARK: - MTKViewDelegate
-
-extension VRMRenderer: MTKViewDelegate {
-    public func mtkView(_ view: MTKView, drawableSizeWillChange size: CGSize) {
-        let aspect = Float(size.width / size.height)
-        projectionMatrix = makePerspective(fovyRadians: .pi / 3, aspectRatio: aspect, nearZ: 0.1, farZ: 100.0)
-    }
-
-    public func draw(in view: MTKView) {
-        guard let commandBuffer = commandQueue.makeCommandBuffer(),
-              let descriptor = view.currentRenderPassDescriptor else {
-            return
-        }
-
-        draw(in: view, commandBuffer: commandBuffer, renderPassDescriptor: descriptor)
-
-        if let drawable = view.currentDrawable {
-            commandBuffer.present(drawable)
-        }
-
-        // Add comprehensive error handler for debugging
-        commandBuffer.addCompletedHandler { [weak self] cb in
-            if cb.status == .error {
-                vrmLog("[VRMRenderer] ❌ METAL ERROR: Command buffer failed!")
-                if let error = cb.error {
-                    vrmLog("[VRMRenderer] ❌ Error details: \(error)")
-                }
-                // Log additional debug info
-                if let model = self?.model {
-                    var totalPrimitives = 0
-                    var maxMorphs = 0
-                    for mesh in model.meshes {
-                        totalPrimitives += mesh.primitives.count
-                        for prim in mesh.primitives {
-                            maxMorphs = max(maxMorphs, prim.morphTargets.count)
-                        }
-                    }
-                    vrmLog("[VRMRenderer] ❌ Model stats: \(totalPrimitives) primitives, max \(maxMorphs) morphs per primitive")
-                }
-            } else if cb.status == .completed {
-                // Success - log occasionally for complex models
-                if let frameCounter = self?.frameCounter, frameCounter % 300 == 0 {
-                    if let model = self?.model {
-                        var totalPrimitives = 0
-                        for mesh in model.meshes {
-                            totalPrimitives += mesh.primitives.count
-                        }
-                        if totalPrimitives > 50 {
-                            vrmLog("[VRMRenderer] ✅ Frame \(frameCounter): Successfully rendered \(totalPrimitives) primitives")
-                        }
-                    }
-                }
-            }
-        }
-
-        commandBuffer.commit()
-    }
-
-    // MARK: - SpringBone Debug Controls
-
-    public func resetSpringBone() {
-        // GPU system resets automatically on model load
-        vrmLog("[VRMRenderer] SpringBone reset (GPU system resets on model load)")
-    }
-
-    public func applySpringBoneForce(gravity: SIMD3<Float>? = nil, wind: SIMD3<Float>? = nil, duration: Float = 1.0) {
-        if let gravity = gravity {
-            temporaryGravity = gravity
-        }
-        if let wind = wind {
-            temporaryWind = wind
-        }
-        forceTimer = duration
-    }
-
-    private func updateSpringBoneForces(deltaTime: Float) {
-        // Apply temporary forces if timer is active
-        if forceTimer > 0 {
-            if let gravity = temporaryGravity {
-                model?.springBoneGlobalParams?.gravity = gravity
-            }
-            if let wind = temporaryWind {
-                model?.springBoneGlobalParams?.windDirection = simd_normalize(wind)
-                model?.springBoneGlobalParams?.windAmplitude = simd_length(wind)
-            }
-            forceTimer -= deltaTime
-        } else if temporaryGravity != nil || temporaryWind != nil {
-            // Timer expired - restore initial gravity and clear wind
-            // Only restore if we had temporary overrides (don't overwrite initial setup)
-            if temporaryGravity != nil {
-                model?.springBoneGlobalParams?.gravity = [0, -9.8, 0]
-            }
-            if temporaryWind != nil {
-                model?.springBoneGlobalParams?.windAmplitude = 0
-            }
-            temporaryGravity = nil
-            temporaryWind = nil
-        }
-    }
 }
